% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinedistance.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Compute cosine distances}
\usage{
cosine_distance(x, y = NULL)
}
\arguments{
\item{x}{A vector or matrix.}

\item{y}{An optional second vector or matrix}
}
\value{
The cosine distance
}
\description{
Cosine distance is simply 1 - cosine_similarity
}
