% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinedistance.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute cosine similarity}
\usage{
cosine_similarity(x, y = NULL)
}
\arguments{
\item{x}{A vector or matrix.}

\item{y}{An optional second vector or matrix}
}
\value{
The cosine similarity
}
\description{
Cosine similarity is the cosine of the angle between the two rays that each
pass through the origin and another point. Cosine is computed as the inner
product of the two vectors divided by the product of their norms.
}
