% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinedistance.R
\name{cossimX}
\alias{cossimX}
\title{Pairwise cosine similarity}
\usage{
cossimX(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
A symmetric matrix of cosine similarity values.
}
\description{
Compare each pair of rows
}
