% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinedistance.R
\name{cossimXYMatVec}
\alias{cossimXYMatVec}
\title{Cosine similarity between multiple rows and single vector}
\usage{
cossimXYMatVec(x, y)
}
\arguments{
\item{x}{Matrix}

\item{y}{Vector}
}
\value{
A similarity value for each row in \code{x}
}
\description{
Cosine similarity between multiple rows and single vector
}
