% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinedistance.R
\name{cossimXYVecVec}
\alias{cossimXYVecVec}
\title{Cosine similarity between a pair of vectors}
\usage{
cossimXYVecVec(x, y)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}
}
\value{
A single similarity value
}
\description{
Vectors \code{x} and \code{y} must be equal length
}
